% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bloque_marque.R
\name{add_bloque_marque}
\alias{add_bloque_marque}
\title{add the bloque_marque to a plot}
\usage{
add_bloque_marque(plot = ggplot2::last_plot(), header, ratio = 10)
}
\arguments{
\item{plot}{the plot}

\item{header}{the header to add}

\item{ratio}{the ratio in \% between the header and the plot}
}
\value{
a ggplot
}
\description{
add the bloque_marque to a plot
}
