% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.R
\name{add_logo}
\alias{add_logo}
\title{Add a logo to a plot}
\usage{
add_logo(plot = ggplot2::last_plot(), header, ratio = 10)
}
\arguments{
\item{plot}{A plot.}

\item{header}{The header to add.}

\item{ratio}{The ratio in \% between the header and the plot.}
}
\value{
A \code{ggplot} object.
}
\description{
Add a logo to a plot
}
