% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemes.R
\name{check_fonts_in_r}
\alias{check_fonts_in_r}
\title{Tests if fonts are available on the system and import for extrafont}
\usage{
check_fonts_in_r(
  fonts = c("Marianne", "Spectral"),
  import = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{fonts}{vector of fonts}

\item{import}{Logical. Whether to import ttf in R. (Otherwise just check)}

\item{verbose}{list fonts to install if needed}
}
\description{
Tests if fonts are available on the system and import for extrafont
}
\details{
For fonts to be used, they need to be installed on the machine.
}
\examples{
check_fonts_in_r(import = FALSE)

# Test if working
\dontrun{
library(ggplot2)
ggplot(mtcars) + geom_point(aes(cyl, mpg)) + labs(title = "Mon titre avec Marianne", 
    x = "lab X avec Spectral", y = "Lab Y avec Spectral") + 
    theme(title = element_text(family = "Marianne"), 
        axis.title.x = element_text(family = "Spectral"), 
        axis.title.y = element_text(family = "Spectral"))
}
}
