% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bloque_marque.R
\name{create_bloque_marque}
\alias{create_bloque_marque}
\title{Create a grob from image file}
\usage{
create_bloque_marque(file = NULL, alias = NULL, x_pos = 0.04, y_pos = 0.96)
}
\arguments{
\item{file}{local file}

\item{alias}{name from a bloque marque available in gouvdown}

\item{x_pos, y_pos}{x and y position of the file relative to the border}
}
\value{
a grob
}
\description{
Create a grob from image file
}
\examples{
\dontrun{
create_bloque_marque(alias = "marianne")
}
}
