% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_gouv.R
\name{create_header_html_gouv}
\alias{create_header_html_gouv}
\title{create html_gouv bootstrap header}
\usage{
create_header_html_gouv(
  logo = NULL,
  file = logo_file_path(logo),
  output = "header.html"
)
}
\arguments{
\item{logo}{Name of a logo available in \code{gouvdown}. The list of available
logos can be obtained with \code{\link[=list_logos]{list_logos()}}.}

\item{file}{Path to a \code{png} file.}

\item{output}{name of the header. The logo is copied in an 'www' subfolder}
}
\value{
an html file
}
\description{
create html_gouv bootstrap header
}
\examples{
create_header_html_gouv(logo = 'gouvernement')
}
