% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.R
\name{create_logo}
\alias{create_logo}
\title{Create a grid graphical object from an image file}
\usage{
create_logo(
  logo = NULL,
  file = logo_file_path(logo),
  x_pos = 0.04,
  y_pos = 0.96
)
}
\arguments{
\item{logo}{Name of a logo available in \code{gouvdown}.}

\item{file}{Path to a local \code{png} file.}

\item{x_pos, y_pos}{\code{x} and \code{y} positions of the logo relative to the border.}
}
\value{
An object of class "grob".
}
\description{
Create a grid graphical object from an image file
}
\details{
Fill either \code{file} or \code{logo} parameter:  an error is thrown if both parameters are provided.
}
\examples{
create_logo("marianne")
}
