% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{display_palette}
\alias{display_palette}
\alias{display_palette_all}
\title{Display palettes used in French governmental documents}
\usage{
display_palette(palette)

display_palette_all()
}
\arguments{
\item{palette}{Character name of palette in \link{gouv_palettes}.}
}
\description{
Plot the palettes used in \code{gouvdown}.
}
\examples{
display_palette("pal_gouv_fr")
display_palette_all()
}
