% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{fonts-dependencies}
\alias{fonts-dependencies}
\alias{marianne_font_dep}
\alias{spectral_font_dep}
\alias{spectral_sc_font_dep}
\title{HTML dependencies for fonts used in French governmental documents}
\usage{
marianne_font_dep()

spectral_font_dep(use_gouvdown_fonts = TRUE)

spectral_sc_font_dep(use_gouvdown_fonts = TRUE)
}
\arguments{
\item{use_gouvdown_fonts}{Do you prefer to use \code{gouvdown.fonts} dependencies?}
}
\value{
An object that can be included in a list of dependencies passed to
\link[htmltools:attachDependencies]{attachDependencies}.
}
\description{
These functions provide common HTML dependencies for re-use by other R
Markdown output formats or Shiny applications.
}
\examples{
spectral_font_dep()
}
\keyword{internal}
