% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloque_marque.R
\name{get_bloque_marque}
\alias{get_bloque_marque}
\title{get the link to the bloque marque in the package directory}
\usage{
get_bloque_marque(bloque_marque)
}
\arguments{
\item{bloque_marque}{}
}
\value{

}
\description{
get the link to the bloque marque in the package directory
}
