% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloque_marque.R
\name{get_logo}
\alias{get_logo}
\title{Get the path to the logo file in the package directory}
\usage{
get_logo(logo)
}
\arguments{
\item{logo}{Name of the logo.}
}
\value{
A character string with the path to the logo file.
}
\description{
Get the path to the logo file in the package directory
}
