% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{gglogo}
\alias{gglogo}
\title{Create a graphical object from a logo file}
\usage{
gglogo(
  logo,
  file = logo_file_path(logo),
  x = 0.04,
  y = 0.96,
  just = c("left", "top"),
  ...
)
}
\arguments{
\item{logo}{Name of a logo available in \code{gouvdown}. The list of available
logos can be obtained with \code{\link[=list_logos]{list_logos()}}.}

\item{file}{Path to a \code{png} file.}

\item{x, y}{\code{x} and \code{y} positions of the logo relative to the border. Since
these arguments are passed to \code{\link[grid:grid.raster]{grid::rasterGrob()}},
\link[grid:unit]{unit objects} are allowed.}

\item{just}{The justification of the rectangle
    relative to its (x, y) location.  If there are two values, the first
    value specifies horizontal justification and the second value specifies
    vertical justification.  Possible string values are: \code{"left"},
    \code{"right"}, \code{"centre"}, \code{"center"}, \code{"bottom"},
    and \code{"top"}.  For numeric values, 0 means left alignment
    and 1 means right alignment.
  }

\item{...}{Other arguments passed to \code{\link[grid:grid.raster]{grid::rasterGrob()}}.}
}
\value{
An object of class "grob".
}
\description{
The \code{gglogo()} function creates a \code{grob} (a \code{grid} graphical object) from
a logo file. This \code{grob} can be combined with a \code{ggplot2} plot object.
}
\details{
Fill either \code{logo} or \code{file} parameter: an error is thrown if both
parameters are provided.
}
\examples{
library(grid)

header <- gglogo("marianne", x = 0, y = 1)
vp <- viewport(x = 0.22, y = 0.78, width = 0.2, height = 0.2, just = c("right", 
    "bottom"))
pushViewport(vp)
grid.draw(header)
}
