% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitbook_gouv.R
\name{gitbook_gouv}
\alias{gitbook_gouv}
\title{The gouvdown GitBook output format}
\usage{
gitbook_gouv(
  fig_caption = TRUE,
  number_sections = TRUE,
  self_contained = FALSE,
  lib_dir = "libs",
  pandoc_args = NULL,
  ...,
  template = "default",
  split_by = c("chapter", "chapter+number", "section", "section+number", "rmd", "none"),
  split_bib = TRUE,
  config = list(),
  table_css = TRUE
)
}
\arguments{
\item{fig_caption, number_sections, self_contained, lib_dir, pandoc_args}{...
Arguments to be passed to \code{rmarkdown::\link{html_document}()}
(\code{...} not including \code{toc}, and \code{theme}).}

\item{template}{Pandoc template to use for rendering. Pass \code{"default"}
to use the bookdown default template; pass a path to use a custom template.
The default template should be sufficient for most use cases. In case you
want to develop a custom template, we highly recommend to start from the
default template:
\url{https://github.com/rstudio/bookdown/blob/master/inst/templates/gitbook.html}.}

\item{config}{A list of configuration options for the gitbook style, such as
the font/theme settings.}

\item{table_css}{\code{TRUE} to load gitbook's default CSS for tables. Choose
\code{FALSE} to unload and use customized CSS (for exmaple, bootstrap) via
the \code{css} option. Default is \code{TRUE}.}
}
\description{
This output format function ported a style provided by GitBook
(\url{https://www.gitbook.com}) for R Markdown.
}
