% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{gouv_pal_inter}
\alias{gouv_pal_inter}
\title{Interpolation des palettes}
\usage{
gouv_pal_inter(palette = "pal_gouv_fr", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in omni_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Interpolation des palettes
}
\keyword{internal}
