% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gouv_palettes.R
\docType{data}
\name{gouv_palettes}
\alias{gouv_palettes}
\title{Colors palettes for France design system}
\format{
A named list with 23 palettes.
}
\usage{
gouv_palettes
}
\description{
Colors palettes for France design system
}
\references{
\url{https://www.gouvernement.fr/charte/charte-graphique-les-fondamentaux/les-couleurs}
}
\keyword{datasets}
