% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gouvdown-package.R
\docType{package}
\name{gouvdown-package}
\alias{gouvdown}
\alias{gouvdown-package}
\title{gouvdown: French Government Design System for R Markdown}
\description{
Produce documents compliants with the French government design system.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/spyrales/gouvdown}
  \item Report bugs at \url{https://github.com/spyrales/gouvdown/issues}
}

}
\author{
\strong{Maintainer}: Maël Theulière \email{mael.theuliere@gmail.com}

Authors:
\itemize{
  \item Romain Lesur \email{romain.lesur@gmail.com} (\href{https://orcid.org/0000-0002-0721-5595}{ORCID})
  \item Thomas Vroylandt \email{tvroylandt@gmail.com}
}

Other contributors:
\itemize{
  \item  République Française [copyright holder]
  \item  ThinkR (helper functions from drealdown) [copyright holder]
}

}
\keyword{internal}
