% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{html_dependency_spectral}
\alias{html_dependency_spectral}
\title{HTML dependencies for French government web documents}
\usage{
html_dependency_spectral()
}
\value{
An object that can be included in a list of dependencies passed to
\link[htmltools:attachDependencies]{attachDependencies}.
}
\description{
These functions provide common HTML dependencies for re-use by other R
Markdown output formats or Shiny applications.
}
\examples{
html_dependency_spectral()
}
\keyword{internal}
