% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_gouv.R
\name{html_gouv}
\alias{html_gouv}
\title{gouvdown html document}
\usage{
html_gouv(
  ...,
  extra_dependencies = list(),
  includes = NULL,
  css = NULL,
  logo = NULL
)
}
\arguments{
\item{...}{Additional arguments passed to \code{rmarkdown::\link{html_document}()}.}

\item{extra_dependencies}{extra dep}

\item{includes}{Named list of additional content to include within the document (typically created using the \code{rmarkdown::\link{includes}()} function).}

\item{css}{css}

\item{logo}{a logo available in \code{gouvdown::\link{list_logos}()} or a file path to image file}
}
\value{
An R Markdown output format object to be passed to
\code{rmarkdown::\link{render}()}.
}
\description{
gouvdown html document
}
