% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_gouv_paged.R
\name{html_paged_long_gouv}
\alias{html_paged_long_gouv}
\title{gouvdown pdf paged document - long version}
\usage{
html_paged_long_gouv(
  ...,
  extra_dependencies = list(),
  css = NULL,
  header = "header.html"
)
}
\arguments{
\item{...}{Additional arguments passed to \code{rmarkdown::\link{html_document}()}.}

\item{extra_dependencies}{extra dep}

\item{css}{css}

\item{header}{header to include}
}
\description{
gouvdown pdf paged document - long version
}
