% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{lettre_word}
\alias{lettre_word}
\title{Word output format for creating a docx letter}
\usage{
lettre_word(reference_docx = pkg_resource("docx", "template_lettre.docx"), ...)
}
\arguments{
\item{reference_docx}{Use the specified file as a style reference in
producing a docx file. For best results, the reference docx should be a
modified version of a docx file produced using pandoc. Pass "default"
to use the rmarkdown default styles.}

\item{...}{Additional arguments passed to \code{rmarkdown::\link{word_document}()}.}
}
\value{
An R Markdown output format object to be passed to
\code{rmarkdown::\link{render}()}.
}
\description{
Word output format for creating a docx letter
}
