% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logo.R
\name{list_logos}
\alias{list_logos}
\title{List available logos}
\usage{
list_logos()
}
\value{
A list with the names of the logos available in \code{gouvdown}.
}
\description{
\code{list_logos()} is a helper function that can be used to obtain the list of
the official design marks contained in the \code{gouvdown} package.
}
