% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logo.R
\name{logo_file_path}
\alias{logo_file_path}
\title{Get the path to the logo file in the package directory}
\usage{
logo_file_path(logo)
}
\arguments{
\item{logo}{Name of the logo.}
}
\value{
A character string with the path to the logo file.
}
\description{
Get the path to the logo file in the package directory
}
