% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propre_brochure.R
\name{propre_brochure}
\alias{propre_brochure}
\title{gouvdown PDF brochure - PROPRE style}
\usage{
propre_brochure(..., css = NULL)
}
\arguments{
\item{...}{Additional arguments passed to \code{bookdown::\link{html_document2}()}.}

\item{css}{A character vector of additionnal CSS file paths.}
}
\value{
An R Markdown output format object to be passed to
\code{rmarkdown::\link{render}()}.
}
\description{
gouvdown PDF brochure - PROPRE style
}
