% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color_gouv_continuous}
\alias{scale_color_gouv_continuous}
\title{Continuous color scale}
\usage{
scale_color_gouv_continuous(palette = "pal_gouv_a", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in \link{gouv_palettes}.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:scale_color_gradientn]{ggplot2::scale_color_gradientn()}}.}
}
\description{
Continuous color scale
}
