% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color_gouv_discrete}
\alias{scale_color_gouv_discrete}
\title{Discrete color scale}
\usage{
scale_color_gouv_discrete(palette = "pal_gouv_fr", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in \link{gouv_palettes}, one of \verb{r paste0("``", names(gouvdown::gouv_palettes), "``", collapse = ", ")}.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
Discrete color scale
}
