% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_gouv_discrete}
\alias{scale_fill_gouv_discrete}
\title{Echelle de fill discrète}
\usage{
scale_fill_gouv_discrete(palette = "pal_gouv_fr", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Nom de la palette}

\item{reverse}{Palette inversée ou non}

\item{...}{Additional arguments}
}
\description{
Echelle de fill discrète
}
