% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemes.R
\name{theme_grey_gouv}
\alias{theme_grey_gouv}
\title{ggplot2 grey theme with french government design template}
\usage{
theme_grey_gouv(...)
}
\arguments{
\item{...}{Other params passed to \code{theme_gouv()}.}
}
\value{
A \code{ggplot2} theme.
}
\description{
ggplot2 grey theme with french government design template
}
\examples{
\dontrun{
library(ggplot2)
ggplot(data = iris, aes(x = Sepal.Length, y = Sepal.Width)) +
  geom_point(aes(color = Species, shape = Species)) +
  labs(
    x = "Sepal Length", y = "Sepal Width",
    title = "Sepal Length-Width",
    subtitle = "with gouv_theme()"
  ) +
  theme_grey_gouv() +
  scale_color_gouv_discrete(palette = "pal_gouv_fr")
}

}
