% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{check_fonts_in_r}
\alias{check_fonts_in_r}
\title{Test if fonts are available on the system and import them in R}
\usage{
check_fonts_in_r(
  fonts = c("Marianne", "Spectral"),
  import = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{fonts}{Vector of fonts.}

\item{import}{Logical. Whether to import \code{ttf} font files in R (otherwise
just check).}

\item{verbose}{List fonts to install if needed.}
}
\description{
Test if fonts are available on the system and import them in R
}
\details{
For fonts to be used, they need to be installed on the machine.
}
\examples{
\donttest{
check_fonts_in_r(import = FALSE)

# Test if working
library(ggplot2)
ggplot(mtcars) +
  geom_point(aes(cyl, mpg)) +
  labs(
    title = "Mon titre avec Marianne",
    x = "lab X avec Spectral",
    y = "Lab Y avec Spectral"
  ) +
  theme(
    title = element_text(family = "Marianne"),
    axis.title.x = element_text(family = "Spectral"),
    axis.title.y = element_text(family = "Spectral")
  )
}
}
