% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{display_palette}
\alias{display_palette}
\alias{display_palette_all}
\title{Display palettes used in French governmental documents}
\usage{
display_palette(palette = "scale_color_gouv_discrete")

display_palette_all()
}
\arguments{
\item{palette}{Character name of palette in \link{gouv_palettes}, one of \code{pal_gouv_fr}, \code{pal_gouv_a}, \code{pal_gouv_b}, \code{pal_gouv_c}, \code{pal_gouv_d}, \code{pal_gouv_e}, \code{pal_gouv_f}, \code{pal_gouv_g}, \code{pal_gouv_h}, \code{pal_gouv_i}, \code{pal_gouv_j}, \code{pal_gouv_k}, \code{pal_gouv_l}, \code{pal_gouv_m}, \code{pal_gouv_n}, \code{pal_gouv_o}, \code{pal_gouv_p}, \code{pal_gouv_q}, \code{pal_gouv_r}, \code{pal_gouv_qual1}, \code{pal_gouv_qual2}, \code{pal_gouv_div1}, \code{pal_gouv_div2}.}
}
\description{
Plot the palettes used in \code{gouvdown}.
}
\examples{
display_palette("pal_gouv_fr")
display_palette_all()
}
