% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitbook_gouv.R
\name{gitbook_gouv}
\alias{gitbook_gouv}
\title{The gouvdown GitBook output format}
\usage{
gitbook_gouv(..., extra_dependencies = list(), use_gouvdown_fonts = TRUE)
}
\arguments{
\item{...}{Additional arguments passed to \code{bookdown::\link{gitbook}()}.}

\item{extra_dependencies}{Additional function arguments to pass to the
base R Markdown HTML output formatter \code{\link[rmarkdown]{html_document_base}}}

\item{use_gouvdown_fonts}{Use the fonts provided by the \code{gouvdown.fonts} package, if installed.}
}
\value{
An R Markdown output format object to be passed to
\code{rmarkdown::\link{render}()}.
}
\description{
The gouvdown GitBook output format
}
