% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{gouv_colors}
\alias{gouv_colors}
\title{Colors vector}
\usage{
gouv_colors(...)
}
\arguments{
\item{...}{The color you want, by code (from a0 to r4 or bleu_france, blanc, rouge_marianne)}
}
\value{

}
\description{
gouv_colors() directly gives the colors as hex codes
}
\examples{
gouv_colors("a1")

}
