% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{gouv_colors}
\alias{gouv_colors}
\title{Obtenir les couleurs de la charte graphique}
\usage{
gouv_colors(...)
}
\arguments{
\item{...}{La couleur souhaitée}
}
\value{

}
\description{
gouv_colors() donne directement les couleurs principales
de la charte graphique
}
\examples{
gouv_colors("a1")

}
