% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{gouv_pal_inter}
\alias{gouv_pal_inter}
\title{Palettes interpolation}
\usage{
gouv_pal_inter(palette = "pal_gouv_fr", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette ingouv_palette}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Palettes interpolation
}
\keyword{internal}
