% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{gouv_pal_inter}
\alias{gouv_pal_inter}
\title{Palettes interpolation}
\usage{
gouv_pal_inter(palette = "pal_gouv_fr", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in \link{gouv_palettes}, one of \verb{r paste0("``", names(gouvdown::gouv_palettes), "``", collapse = ", ")}.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments passed to \code{\link[grDevices:colorRampPalette]{grDevices::colorRampPalette()}}.}
}
\value{
A function that takes an integer argument (the required number of
colors) and returns a character vector of colors (see \link[grDevices:rgb]{grDevices::rgb})
interpolating the given sequence.
}
\description{
Palettes interpolation
}
\keyword{internal}
