% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gouv_palettes.R
\docType{data}
\name{gouv_palettes}
\alias{gouv_palettes}
\title{Colors palettes for France design system}
\format{
A named list with 23 palettes.
}
\usage{
gouv_palettes
}
\description{
\if{html}{\out{<div class="sourceCode {r palettes, echo=FALSE}">}}\preformatted{pal <- gouvdown::gouv_palettes
labels <- names(pal)
do.call(pals::pal.bands,
        c(pal, list(labels = labels, gap = 0.2, main = "gouvdown palettes"))
)
}\if{html}{\out{</div>}}
}
\references{
\url{https://www.gouvernement.fr/charte/charte-graphique-les-fondamentaux/les-couleurs}
}
\keyword{datasets}
