% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_gouv.R
\name{html_gouv}
\alias{html_gouv}
\title{gouvdown html document}
\usage{
html_gouv(..., extra_dependencies = list(), css = NULL)
}
\arguments{
\item{extra_dependencies}{extra dep}

\item{css}{csss}
}
\value{

}
\description{
gouvdown html document
}
