% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_gouv.R
\name{html_gouv}
\alias{html_gouv}
\title{gouvdown HTML document}
\usage{
html_gouv(
  extra_dependencies = list(),
  includes = NULL,
  ...,
  logo = NULL,
  use_gouvdown_fonts = TRUE
)
}
\arguments{
\item{extra_dependencies}{Extra dependencies as a list of the
\code{html_dependency} class objects typically generated by
\code{htmltools::\link{htmlDependency}()}.}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link[rmarkdown]{includes}} function).}

\item{...}{Additional arguments passed to \code{rmarkdown::\link{html_document}()}.}

\item{logo}{a logo available in \code{gouvdown::\link{list_logos}()} or a file path to image file}

\item{use_gouvdown_fonts}{Use the fonts provided by the \code{gouvdown.fonts} package, if installed.}
}
\value{
An R Markdown output format object to be passed to
\code{rmarkdown::\link{render}()}.
}
\description{
gouvdown HTML document
}
