% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propre_brochure.R
\name{propre_brochure}
\alias{propre_brochure}
\title{gouvdown PDF brochure - PROPRE style}
\usage{
propre_brochure(
  extra_dependencies = list(),
  includes = NULL,
  ...,
  logo = NULL,
  css = NULL,
  use_gouvdown_fonts = TRUE,
  width_main_column = 15,
  made_with = "Ce document a été réalisé avec le package `gouvdown`"
)
}
\arguments{
\item{extra_dependencies}{Extra dependencies as a list of the
\code{html_dependency} class objects typically generated by
\code{htmltools::\link{htmlDependency}()}.}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link[rmarkdown]{includes}} function).}

\item{...}{Additional arguments passed to \code{bookdown::\link{html_document2}()}.}

\item{logo}{a logo available in \code{gouvdown::\link{list_logos}()} or a file path to image file}

\item{css}{A character vector of additionnal CSS file paths.}

\item{use_gouvdown_fonts}{Use the fonts provided by the \code{gouvdown.fonts} package, if installed.}

\item{width_main_column}{Width of the main text column (in \%)}

\item{made_with}{Character. Add extra information to show how the document was made.}
}
\value{
An R Markdown output format object to be passed to
\code{rmarkdown::\link{render}()}.
}
\description{
gouvdown PDF brochure - PROPRE style
}
