% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color_gouv_continuous}
\alias{scale_color_gouv_continuous}
\title{Echelle de couleur continue}
\usage{
scale_color_gouv_continuous(palette = "pal_gouv_a", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Nom de la palette}

\item{reverse}{Palette inversée ou non}

\item{...}{Additional arguments}
}
\description{
Echelle de couleur continue
}
