% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color_gouv_discrete}
\alias{scale_color_gouv_discrete}
\title{Discrete color scale}
\usage{
scale_color_gouv_discrete(palette = "pal_gouv_fr", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Palette name}

\item{reverse}{Reverse the palette or not}

\item{...}{Additional arguments}
}
\description{
Discrete color scale
}
