% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_gouv_continuous}
\alias{scale_fill_gouv_continuous}
\title{Continuous fill scale}
\usage{
scale_fill_gouv_continuous(palette = "pal_gouv_a", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Palette name}

\item{reverse}{Reverse the palette or not}

\item{...}{Additional arguments}
}
\description{
Continuous fill scale
}
