% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_gouv_continuous}
\alias{scale_fill_gouv_continuous}
\title{Continuous fill scale}
\usage{
scale_fill_gouv_continuous(palette = "pal_gouv_a", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in \link{gouv_palettes}, one of \verb{r paste0("``", names(gouvdown::gouv_palettes), "``", collapse = ", ")}.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:scale_fill_gradientn]{ggplot2::scale_fill_gradientn()}}.}
}
\description{
Continuous fill scale
}
