% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_gouv_discrete}
\alias{scale_fill_gouv_discrete}
\title{Discrete fill scale}
\usage{
scale_fill_gouv_discrete(palette = "pal_gouv_fr", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in \link{gouv_palettes}, one of \verb{r paste0("``", names(gouvdown::gouv_palettes), "``", collapse = ", ")}.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{
  Arguments passed on to \code{\link[=scale_color_gouv_discrete]{scale_color_gouv_discrete}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Discrete fill scale
}
