% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemes.R
\name{theme_gouv}
\alias{theme_gouv}
\title{ggplot2 theme with french government design template}
\usage{
theme_gouv(
  base_family = "Spectral",
  base_size = 12,
  plot_title_family = "Marianne",
  plot_title_size = 28,
  plot_title_face = "bold",
  plot_title_margin = 10,
  subtitle_family = "Marianne Light",
  subtitle_size = 22,
  subtitle_face = "plain",
  subtitle_margin = 15,
  strip_text_family = base_family,
  strip_text_size = 22,
  strip_text_face = "plain",
  caption_family = base_family,
  caption_size = 12,
  caption_face = "plain",
  caption_margin = 10,
  axis_text_size = base_size,
  axis_title_family = subtitle_family,
  axis_title_size = base_size,
  axis_title_face = "plain",
  axis_title_just = "rt",
  plot_margin = margin(30, 30, 30, 30),
  grid_col = "#888888",
  grid = TRUE,
  panel_background_color = "#ffffff",
  panel_border_color = "#ffffff",
  legend_background_color = "#ffffff",
  legend_border_color = "#ffffff",
  axis_col = "#cccccc",
  axis = FALSE,
  ticks = FALSE
)
}
\arguments{
\item{base_family, base_size}{Base font family and size.}

\item{plot_title_family, plot_title_face, plot_title_size, plot_title_margin}{Plot title family, face, size and margin.}

\item{subtitle_family, subtitle_face, subtitle_size}{Plot subtitle family, face and size.}

\item{subtitle_margin}{Plot subtitle margin bottom (single numeric value).}

\item{strip_text_family, strip_text_face, strip_text_size}{Facet label font family, face and size.}

\item{caption_family, caption_face, caption_size, caption_margin}{Plot caption family, face, size and margin.}

\item{axis_text_size}{Font size of axis text.}

\item{axis_title_family, axis_title_face, axis_title_size}{Axis title font family, face and size.}

\item{axis_title_just}{Axis title font justification, one of \verb{[blmcrt]}.}

\item{plot_margin}{Plot margin (specify with \code{ggplot2::margin()}).}

\item{grid_col, axis_col}{Grid and axis colors; defaults: \verb{#888888} for grid color and \verb{#CCCCCC} for axis color.}

\item{grid}{Panel grid (\code{TRUE}, \code{FALSE} or a combination of \code{X}, \code{x}, \code{Y}, \code{y}).}

\item{panel_background_color}{Panel background fill color; default to \code{white}.}

\item{panel_border_color}{Panel border color; default to \code{white}.}

\item{legend_background_color}{Legend background fill color; default to \code{white}.}

\item{legend_border_color}{Legend border color; default to \code{white}.}

\item{axis}{Add x or y axes? \code{TRUE}, \code{FALSE} or "\code{xy}".}

\item{ticks}{Ticks: if \code{TRUE}, add ticks.}
}
\value{
A \code{ggplot2} theme.
}
\description{
ggplot2 theme with french government design template
}
\details{
All font sizes are set in points and colors are either a color name or a hex code.
}
\examples{
\dontrun{
library(ggplot2)
ggplot(data = iris, aes(x = Sepal.Length, y = Sepal.Width)) +
  geom_point(aes(color = Species, shape = Species)) +
  labs(
    x = "Sepal Length", y = "Sepal Width",
    title = "Sepal Length-Width",
    subtitle = "with theme_gouv()"
  ) +
  theme_gouv() +
  scale_color_gouv_discrete()
}

}
